mv queue stuff to its own queue.go file
