+++ satellite/audit/reservoirchore.go
@@ -17,51 +17,6 @@ import (
 	"storj.io/storj/satellite/metainfo"
 )
 
-// queue is a list of paths to audit, shared between the reservoir chore and audit workers.
-type queue struct {
-	cond   sync.Cond
-	queue  []storj.Path
-	closed chan struct{}
-}
-
-func newQueue(cond sync.Cond, closed chan struct{}) *queue {
-	return &queue{
-		cond:   cond,
-		closed: closed,
-	}
-}
-
-// swap switches the backing queue slice with a new queue slice.
-func (queue *queue) swap(newQueue []storj.Path) {
-	queue.cond.L.Lock()
-	queue.queue = newQueue
-	// Notify workers that queue has been repopulated.
-	queue.cond.Broadcast()
-	queue.cond.L.Unlock()
-}
-
-// next gets the next item in the queue.
-func (queue *queue) next(ctx context.Context) (storj.Path, error) {
-	queue.cond.L.Lock()
-	defer queue.cond.L.Unlock()
-
-	// This waits until the queue is repopulated, closed, or context is canceled.
-	for len(queue.queue) == 0 {
-		select {
-		case <-queue.closed:
-			return "", Error.New("queue is closed")
-		case <-ctx.Done():
-			return "", ctx.Err()
-		default:
-			queue.cond.Wait()
-		}
-	}
-	next := queue.queue[0]
-	queue.queue = queue.queue[1:]
-
-	return next, nil
-}
-
 // ReservoirChore populates reservoirs and the audit queue.
 type ReservoirChore struct {
 	log    *zap.Logger
